Refactor 232 project form component (#245)
* refactor(web): add split components

* refactor(web): join split components

* refactor(web): remove unnecessary code

* chore: doc upt

---------
