+++ apps/web/app/create-project/page.tsx
@@ -1,4 +1,4 @@
-import ProjectForm from '~/components/sections/project/project-form'
+import { ProjectForm } from '~/components/sections/project/project-form'
 
 export default function TestPage() {
 	return (
